electron-test-runner/Test-Foundry/GENAI_LLMJudge_ResponseQuality_Framework/requirements.txt
# Core dependencies
pandas==2.1.0
openpyxl==3.1.2
pyyaml==6.0.1
openai==0.28.1

# Async support
aiohttp==3.9.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1

# Type checking
mypy==1.5.1
types-PyYAML==6.0.12.11
pandas-stubs==2.0.3.230814